pyproject.toml
[tool.poetry]
name = "webestateapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
uvicorn = "^0.27.1"
pydantic = {extras = ["email"], version = "^2.6.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
